Move encoder/decoder from message.go to codec.go
